:hammer: move camera to components
