+++ src/Components/Camera.js
